refactoring post card
