+++ core/ui/notifications/src/main/java/social/firefly/core/ui/notifications/cards/Repost.kt
@@ -9,8 +9,8 @@ import social.firefly.core.ui.notifications.NotificationInteractionsNoOp
 import social.firefly.core.ui.notifications.NotificationUiState
 import social.firefly.core.ui.postcard.PostCardInteractions
 import social.firefly.core.ui.postcard.PostCardInteractionsNoOp
-import social.firefly.core.ui.postcard.PostContent
 import social.firefly.core.ui.postcard.PostContentUiState
+import social.firefly.core.ui.postcard.components.PostContent
 
 @Composable
 internal fun RepostNotificationContent(
